EXAMPLES/zep_examples/krypton-prototype/requirements.txt
zep-cloud
python-dotenv
pydantic